fix: handle several rest api bugs in students

+++ rest-api/routes/api.php
@@ -4,6 +4,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AnimalController;
 use App\Http\Controllers\StudentController;
+use App\Http\Controllers\AuthController;
 
 /*
 |--------------------------------------------------------------------------
@@ -39,11 +40,19 @@ Route::delete('/animals/{id}',[AnimalController::class,'destroy']);
 // method get
 Route::get('/students',[StudentController::class,'index']);
 
-// method post
-Route::post('/students',[StudentController::class,'store']);
+// method post (auth)
+Route::post('/students',[StudentController::class,'store'])->middleware('auth:sanctum');
 
-//method put
-Route::put('/students/{id}',[StudentController::class,'update']);
+//method put (auth)
+Route::put('/students/{id}',[StudentController::class,'update'])->middleware('auth:sanctum');
 
-// method delete
-Route::delete('/students/{id}',[StudentController::class,'destroy']);
+// method delete (auth)
+Route::delete('/students/{id}',[StudentController::class,'destroy'])->middleware('auth:sanctum');
+
+//method get detail
+Route::get('/students/{id}',[StudentController::class,'show']);
+
+
+//Auth Routes
+Route::post('/auth/register', [AuthController::class, 'createUser'])->name('register');
+Route::post('/auth/login', [AuthController::class, 'loginUser'])->name('login');
